refactor(structure): split ops and views into per-family packages, release 1.0.1
Replace the ops.py and notebook.py monoliths with two parallel packages
split by operation family, so einsum is no longer the only separated
operation.

- ops/ holds the pure shape maths per family (reshaping, reductions,
  combining, einsum); the shared _check_axis primitive moves to shape.py
- views/ holds the SVG display layer per family (shapes, reshaping,
  reductions, combining, einsum); einsum.py moves in unchanged
- shared view helpers (_shape_caption_parts, _operand_tint, tint ramps)
  move into visual.py
- tests follow the same family layout
- the public API is unchanged; every existing import keeps resolving
